+++ scripts/bump-npm.sh
@@ -1,0 +1,12 @@
+#!/usr/bin/env bash
+set -eu
+
+pnpm changeset || true
+pnpm changeset version
+
+for pkg in $(ls -d packages/*); do
+    CHANGELOG=$(cat ./$pkg/CHANGELOG.md) envsubst < ./$pkg/README.tmpl.md > ./$pkg/README.md
+    git add ./$pkg/README.md
+done
+
+git commit -am "Merge CHANGELOG into README"