+++ packages/emotion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swc/plugin-emotion
 
+## 11.1.0
+
+### Minor Changes
+
+- a872100: build: Update swc_core to v38
+
 ## 11.0.3
 
 ### Patch Changes