+++ packages/swc-magic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swc/plugin-swc-magic
 
+## 9.1.0
+
+### Minor Changes
+
+- a872100: build: Update swc_core to v38
+
 ## 9.0.4
 
 ### Patch Changes